remove npm dependency from todomvc
