+++ examples/todomvc/src/main.rs
@@ -2,12 +2,12 @@ use dip::{bevy::log::LogPlugin, prelude::*};
 use todomvc::TodoMVCPlugin;
 
 fn main() {
+    let tailwind_css = std::fs::read_to_string("examples/todomvc/public/globals.css")
+        .expect("Cannot find compiled Tailwind CSS file");
+
     App::new()
         .insert_non_send_resource(DesktopSettings::<NoRootProps> {
-            custom_head: Some(format!(
-                "<style>{}</style>",
-                include_str!("../public/main.css")
-            )),
+            custom_head: Some(format!("<style>{tailwind_css}</style>")),
             ..Default::default()
         })
         .add_plugin(TodoMVCPlugin)
