+++ src/plugin.rs
@@ -1,4 +1,5 @@
 mod cli;
+mod dip;
 mod tool;
 
-pub use crate::plugin::{cli::*, tool::*};
+pub use crate::plugin::{cli::*, dip::*, tool::*};
